+++ SpaceInvaders/SpaceInvaders/clock.cpp
@@ -42,11 +42,11 @@ CClock::Initialise()
 {
 	__int64 _TimerFrequency, _currTime;
 	QueryPerformanceFrequency((LARGE_INTEGER*)&_TimerFrequency);
-	m_SecondsPerCount = 1.0f / static_cast<double>(_TimerFrequency);
+	m_SecondsPerCount = 1.0f / static_cast<float>(_TimerFrequency);
 
 	QueryPerformanceCounter((LARGE_INTEGER*)&_currTime);
-	m_fCurrentTime = static_cast<double>(_currTime);
-	m_fLastTime = static_cast<double>(_currTime);
+	m_fCurrentTime = static_cast<float>(_currTime);
+	m_fLastTime = static_cast<float>(_currTime);
 
 	return (true);
 }
@@ -60,13 +60,13 @@ CClock::Process()
 
 	QueryPerformanceCounter((LARGE_INTEGER*)&currTime);
 
-	m_fCurrentTime = static_cast<double>(currTime);
+	m_fCurrentTime = static_cast<float>(currTime);
 
 	//Time difference between this frame and the previous frame
-	m_fDeltaTime = (m_fCurrentTime - m_fLastTime)*m_SecondsPerCount;
+	m_fDeltaTime = static_cast<float>((m_fCurrentTime - m_fLastTime)*m_SecondsPerCount);
 
 	//Prepare for the next frame
-	m_fLastTime = m_fCurrentTime;
+	m_fLastTime = static_cast<float>(m_fCurrentTime);
 
 	//Force non-negative
 	if (m_fDeltaTime < 0.0)
